fix(comment-flow): author to comment linking
Before this commit the `hostabee-comment` was crashing when the author
of a comment defined but an object with a "fullname" nor a "name". Also,
the `hostabee-comment-flow` was losing the "author" property value when
no user was matching to it.

Now the `hostabee-comment` handles the case when author has no (full)name
property and the `hostabee-comment-flow` do not delete the value of
comment.author when this one can't be substitute.