feat(comment flow): add support for mapper
Before this commit the `hostabee-comment-flow` was not able to use
elements inheriting the `hostabee-comment-mapper-mixin`. Now, the
comment flow can use a mapper to make users data model matching to the
model used internally. So users only have to provide mapping
(description of transformation to apply on comments and users objects)
to make the element compatible with their data structure.