chore(demo): add demo for hostabee-comment-flow with custom mapper
Before this commit there was not any demo page to show how to create our
custom mapper and use it with the `hostabee-comment-flow`. Now there is
one and users can see how much easy it is to map their data structure
with the format expected by the `hostabee-comment-flow`. Mappings can
aslo be defined programmatically.