chore(test): remove unrelevant tests
Since the version 1.0.1 of the `hostabee-element` the elements
inheriting from it does not fallback to English if the target language
is not available. So unit tests related to this feature had to be
removed.

See https://github.com/Hostabee/hostabee-element/commit/dd4654082dd585d8969c47f5bb4c433e0b24ec60