feat(form): display file picker button
Before this commit the button to open a file picker was displayed only
when the form was allowing to attach files AND the other buttons (Cancel
and Save) were displyed too.

This commit allow the user to display the file picker button only when
the text area is not empty and even if the other buttons (Cancel and
Save) are hidden.