fix(form): escape key down event
Before the introduction of "cancel" event dispatched when the
cancel button of the form is clicked, the form was simply
cleaned when the user pushed the Escape key. When the
dispatching of "Cancel" event was added the behavior of the
Escape key was impacted. This commit brings the previous
behavior back. Which consists to clear the form when Escape
key is down without dispatching any event.