feat(form): input text area required only when minlegth set
Before this commit the text area input was always required. Since a
comment can now have attachments, some users could want to allow send a
comment wuith pictures only. To do so, the minlength can now to be set
to 0 (or negative value) so the text area does not require a value.